Refactor EditorPane into editor/ directory with macro DnD and TS project refs
- Split EditorPane.tsx (1016 lines) into components/editor/:
  index.tsx (shell), MicroEditorView, MacroView, EditorEmptyState,
  ChapterCard, 5 Lexical plugins, and utils (ids, nodeKeyMapping, structureBuilder)
- Add DragHandlePlugin: micro-view paragraph DnD via @dnd-kit/sortable,
  portal-rendered handles with ResizeObserver-based position tracking
- Add MacroView: chapter-level DnD with @dnd-kit, line-clamp-5 paragraph preview
- Add reorderChapters action to store + reorderChaptersInDocument to documentOps
- Extract reorderByKey to renderer/utils/arrayUtils.ts (shared with AnalysisPane)
- Fix TypeScript project references: add composite+declarationMap to shared/tsconfig,
  add references to tsconfig.renderer/main/preload (resolves VSCode import errors)
- Add *.tsbuildinfo to .gitignore
