+++ .gitignore
@@ -7,3 +7,4 @@ playwright-report
 *.log
 apps/api/data
 .pnpm-store
+*.tsbuildinfo
