+++ apps/desktop/tsconfig.preload.json
@@ -7,5 +7,8 @@
     "declaration": false
   },
   "include": ["src/preload/**/*.ts", "src/preload/**/*.cts"],
-  "exclude": ["src/preload/**/*.test.ts"]
+  "exclude": ["src/preload/**/*.test.ts"],
+  "references": [
+    { "path": "../../packages/shared" }
+  ]
 }
