+++ apps/desktop/tsconfig.renderer.json
@@ -5,5 +5,8 @@
     "types": ["vite/client"]
   },
   "include": ["src/renderer/**/*.ts", "src/renderer/**/*.tsx", "src/renderer/**/*.d.ts"],
-  "exclude": ["src/renderer/**/*.test.ts", "src/main/**/*.test.ts"]
+  "exclude": ["src/renderer/**/*.test.ts", "src/main/**/*.test.ts"],
+  "references": [
+    { "path": "../../packages/shared" }
+  ]
 }
