+++ docs/LiteLizard_spec_v002_old.md
@@ -2,7 +2,7 @@
 
 最終更新: 2026-02-17
 
-このドキュメントは今後の実装における優先仕様とし、既存の v001（旧仕様）より優先する。
+このドキュメントは参照しないでください。
 
 ## 1. 方針
 - 文書本体は Markdown (`.md`) として保存する。
