+++ packages/shared/tsconfig.json
@@ -2,7 +2,9 @@
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "dist",
-    "rootDir": "src"
+    "rootDir": "src",
+    "composite": true,
+    "declarationMap": true
   },
   "include": ["src/**/*.ts", "../../LiteLizard_schema_v1.json"],
   "exclude": ["src/**/*.test.ts"]
